Django/RecipeAppApi/requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4
flake8>=3.6.0,<3.7.0
python-decouple==3.1
django-simple-sso==0.14.1
psycopg2-binary
